Move formatting tests into their own folder, for better organization
