+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Formatting_NetFx/DocumentFormattingTest.cs
@@ -11,7 +11,7 @@ using Xunit;
 using Xunit.Abstractions;
 
 #if COHOSTING
-namespace Microsoft.VisualStudio.Razor.LanguageClient.Cohost;
+namespace Microsoft.VisualStudio.Razor.LanguageClient.Cohost.Formatting;
 #else
 namespace Microsoft.AspNetCore.Razor.LanguageServer.Formatting;
 #endif
